Added formation shift buttons to "Other servers"
Just added them also to limited since I could replace the weird
functions with propper ones that Works (at least for me).
Also added formationRestore and detCityPreArmyUnits functions to the
limitedCreateTweak.
Tested on my server and works BUT mine is not a "testing" server SO I
would appreciate if someone would be so kind as to test this for me
please...